refactor(#3479): extract tui_prompt_relay idle-transcript-scan+rehydration submodules
Phase-2 rank-10 behavior-preserving extraction. PURE MOVE, zero logic change.

- idle_transcript_scan.rs (278 LoC): the pure Claude/Codex transcript+rollout
  prompt scanners (ClaudeIdleTranscriptScan / CodexIdleRolloutScan enums, the
  three scan_* byte-stream parsers, the two *_idle_prompt_observation_should_
  tail_response predicates). No SharedData/http/async-IO coupling.
- rehydration.rs (295 LoC): the Discord-IO/SharedData-coupled Claude TUI binding
  rehydration + dead/orphaned-session eviction pass (rehydrate_existing_claude_
  tui_bindings, evict_dead_orphaned_claude_tui_mirrors, claude_tui_session_is_
  dead_orphaned, pane_is_confirmed_dead_orphaned, rehydrated_claude_tui_binding_
  for_tmux_session). Free fns taking &Arc<SharedData> explicitly, deps reached
  via `use super::*;`, so the move is behavior-identical.

Both modules use `use super::*;`; moved items are pub(super) and re-imported via
`use self::<mod>::{...}` so every call site (and the #[cfg(test)] mod tests
block) stays byte-identical. The sibling helpers shared with non-rehydration
code (claude_tui_runtime_binding_matches_launch, resolve_rehydrated_claude_tmux_
channel_id, parse_claude_tui_launch_script, claude_tui_rehydrate_start_offset,
ClaudeTuiLaunchInfo, the DEAD_ORPHANED_PANE_PROBE_* consts) STAY in root.

tui_prompt_relay.rs prod LoC 5142 -> 4630 (-512). Ratchet baseline lowered to
4630; change-surfaces.md synced; ARCHITECTURE.md inventory regenerated.
